TargaParser: 24bpp almost works! And major rework of decoding
The decoding was split into two functions: tgaDecode and tgaTranslate. The former decodes the Targa image from RLE into RAW without modifying the color data, then the latter translates that color data (if needed) and copies it to the BufferedImage 'image'.
Note: the output of the parseBytes() function is the same format as before; a BufferedImage object.

Long story short: these two functions have to be separate because HzMod's 24bpp Targa images are all mistakenly encoded at 32bpp. :(