Fix memory leak with hanging SwingVideo references
Could actually leak significant amounts of memory through very indirect `Frame` references so it's nice to have this fixed.

Seems to introduce a weird layout bug where the layout is occasionally wrong. This happens only occasionally and is easily fixed by closing and re-opening the window, so I'm leaving it in for now.