Fixed behavior of non-default port for NTR-HR
Port for TCP communication is probably always 8000, so the new code assumes this.
Port for UDP communication is what actually matters, and is configurable via the "Port" option in the NTR-HR menu on the 3DS. Default UDP port is 8001.
So the "Port" setting in Chokistream now corresponds to UDP port rather than TCP port, and the default value of that setting is now 8001.

(Note: the ports are constant in NTR, as far as I'm aware. TCP and UDP ports are always 8000 and 8001 respectively.)