Fix: Added a timeout to NTR UDP socket
This fixes an issue where the thread could get softlocked if it stopped receiving UDP packets, and could not be forcibly stopped.
That became an issue after 710df6e, in which close() was changed to not close the socket.
I could have basically just reverted that change, but I think this is cleaner.